Always use PR number for split post workflow from uploaded metadata
Fixes #82

For some reason the event payload (sometimes?) has an empty array of
`pull_requests`, but due to the way the Action is set up this passes
an empty string to the input `pr_number`. This  causes an error when
trying to convert to `int` to compare to the PR number in the
metadata.