review: Fix split workflow with comments
There is an edge case in split workflow where the `review` does not create
a `clang-tidy-review-output.json` but the rest of the workflow expects the
file to be created even if there are no comments.

This causes `cp` or `mv` operations to fail and the `post` portion of the
workflow fails trying to open an expected file.

This change makes sure the output is created even if there is nothing to
report. The body also serves to debug why there are no comments.