Fix line filter on for clang compiler on windows
The line filters seem to be sensitive to the compiler used.
When using MSVC, the line filter must use backslashes in the name.
However, if the clang compiler is used on Windows, it must be forward.

The simplest solution to this issue is to use an extra entry of the line filter
one with forward slashes and one without.
This allows all compilers to be happy regardless of OS and without looking at
compilation databases.